requirements.txt
APScheduler==3.11.0
asyncmy==0.2.10
DateTime==5.5
fastapi[all]==0.115.8
loguru==0.7.3
openpyxl==3.1.5
pandas==2.2.3
passlib[bcrypt]==1.7.4
Pillow==11.1.0
psutil==7.0.0
pydantic-validation-decorator==0.1.4
PyJWT[crypto]==2.10.1
PyMySQL==1.1.1
redis==5.2.1
requests==2.32.3
SQLAlchemy[asyncio]==2.0.38
user-agents==2.2.0

# 测试
pytest 
pytest-html 
pytest-xdist 
pytest-rerunfailures 
pytest-order 
pytest-result-log 
allure-pytest
pytest_asyncio
anyio
pytest-asyncio
pytest-tornasync
pytest-trio
pytest-twisted
twisted
asgi_lifespan